+++ Services/IApplicationResourceHandler.cs
@@ -1,0 +1,10 @@
+using System.Windows;
+
+namespace SwitchBlade.Services
+{
+    public interface IApplicationResourceHandler
+    {
+        void AddMergedDictionary(ResourceDictionary dictionary);
+        void RemoveMergedDictionary(ResourceDictionary dictionary);
+    }
+}