+++ Services/IDispatcherService.cs
@@ -5,5 +5,6 @@ namespace SwitchBlade.Services
     public interface IDispatcherService
     {
         void Invoke(Action action);
+        System.Threading.Tasks.Task InvokeAsync(Func<System.Threading.Tasks.Task> action);
     }
 }